backend/requirements.txt
flask
python-dotenv
flask-cors
supabase